prevent other user to delete or update posts

+++ controllers/feed.js
@@ -117,6 +117,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized');
+        error.statusCode = 403;
+        throw error;        
+      }
       
       req.file ? imageUrl = req.file.path :  imageUrl = post.imageUrl;
       if (imageUrl !== post.imageUrl) { clearImage(post.imageUrl); }
@@ -145,6 +150,11 @@ exports.deletePost = (req, res, next) => {
         const error = new Error('Could not find a post');
         error.statusCode = 404;
         throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized');
+        error.statusCode = 403;
+        throw error;        
       }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
